Update kcite.php to fix php8 issues, avoid warnings and add a new 'ref' option to the short code
PHP8 requires stricter delineation between static and non-static calls. Replaced all "__CLASS__" statements in the the add_XXXX calls with $this to follow standards. Additionally, the main call is rewritten for the same reason. 

New if tests to avoid code being exercised if variables are not defined. Not quite sure that this is the correct fix though.

New functionality in the shortcode. If [code ref="Jones et al. (2025)"]DOI[/code] is used, the link in the text will use "Jones et al (2025)" as the link anchor. This is useful for in-text citations - note that you can only have one distinct ref in the post.